\name{ERR188297}
\docType{data}
\alias{ERR188}
\alias{ERR188}
\alias{ERR188}
\title{GAlignmentPairs for four samples from GEUVADIS project}
\description{ }
\usage{
data("ERR188297")
}
\format{GAlignmentPairs}
\source{GEUVADIS}
\references{
  Lappalainen, T., et al. Transcriptome and genome sequencing uncovers
  functional variation in humans. Nature. 501(7468): 506–511 (2013)
  doi: 10.1038/nature12531
}
\keyword{datasets}